requirements.txt
flask
PyPDF2
sentence-transformers
faiss-cpu
numpy
pickle5
transformers
